refactor(gba): split bus module implementation
