+++ architecture.md
@@ -281,7 +281,12 @@ All Game Boy Advance hardware lives under `src/gba/`. The module currently provi
 | `src/gba/cpu/arm.rs` | ARM 32-bit instruction decoder/executor. Covers data processing, branch (B/BL), branch-and-exchange (BX), single-data transfer (LDR/STR/LDRB/STRB) and SWI; honours conditional execution. |
 | `src/gba/cpu/thumb.rs` | Thumb 16-bit instruction decoder/executor. Covers move-shifted register, add/subtract, MOV/CMP/ADD/SUB immediate, hi-register operations, BX, PC-relative load, PUSH/POP and conditional/unconditional branches. |
 | `src/gba/cpu/arm7tdmi.rs` | `Arm7tdmi` — fetch/decode/execute pipeline, S/N cycle accounting, exception vectors, and IRQ/FIQ/SWI dispatch (banked register handling, CPSR→SPSR save, vector jump). |
-| `src/gba/bus/mod.rs` | `GbaBus` — system memory bus implementing the CPU `Bus` trait with full GBA address-space routing (BIOS, EWRAM, IWRAM, I/O, PRAM, VRAM, OAM, cart ROM/SRAM), open-bus modelling, BIOS lock, VRAM mirroring, N/S wait-state cycle stubs, and save-state capture/restore for bus-owned memory/peripheral, PPU, APU, and cartridge save-backend state. |
+| `src/gba/bus/mod.rs` | GBA bus module root. Keeps declarations/re-exports minimal and delegates implementation to focused sibling files. |
+| `src/gba/bus/gba_bus.rs` | `GbaBus` storage and inherent API: memory/peripheral ownership, BIOS/ROM loading, stepping peripherals, DMA trigger hooks, save-state capture/restore, open-bus helpers, and mGBA debug console state. |
+| `src/gba/bus/cpu_bus.rs` | CPU `Bus` trait implementation for `GbaBus`, including full GBA address-space routing, read/write side effects, WAITCNT updates, HALTCNT handling, and Game Pak prefetch timing helpers. |
+| `src/gba/bus/dma_bus.rs` | `DmaBus` implementation for `GbaBus`, preserving DMA-specific latch/open-bus behavior while routing transfers through the standard bus paths. |
+| `src/gba/bus/waitstates.rs` | `Waitstates` and `WidthClass`, including WAITCNT-derived N/S cycle lookup tables for BIOS/RAM/I/O/video/Game Pak regions. |
+| `src/gba/bus/addressing.rs` | Shared bus address helpers for VRAM mirroring, timer/DMA control register decoding, Game Pak wait-state predicates, and no-cart open-bus values. |
 | `src/gba/bus/memory.rs` | Backing-store sizes (BIOS/EWRAM/IWRAM/PRAM/VRAM/OAM/SRAM) and helpers for little-endian halfword/word access with mirrored offsets. |
 | `src/gba/bus/io.rs` | `IoRegisters` — dispatch table for the `0x0400_0000`–`0x0400_03FF` I/O window. Routes interrupt-controller and timer registers to live state and provides a backing store for the remaining ~300 registers so unimplemented PPU/APU/DMA registers don't panic. |
 | `src/gba/bus/interrupt.rs` | `InterruptController` — `IE`/`IF`/`IME` registers with write-1-to-clear `IF` semantics and an `irq_line()` predicate consumed by the CPU. |
